feat: credential and network tests

+++ src/services/credential-service/index.ts
@@ -1,16 +1,19 @@
-import { unauthorizedError } from "@/errors";
+import { notFoundError, unauthorizedError } from "@/errors";
 import credentialRepository from "@/repositories/credential-repository";
 import Cryptr from "cryptr";
 
 import { credentialsBody, postingCredential } from "@/utils/credentialUtils";
-import { title } from "process";
+
 import { passwordCryptr } from "@/utils/cryptrUtil";
+import { title } from "process";
 
 async function newCredential(userId: number, credential: credentialsBody) {
-  await credentialRepository.uniqueTitle(userId, title);
+  const titleCheck = await credentialRepository.uniqueTitle(userId, title);
+  if (titleCheck) {
+    throw unauthorizedError();
+  }
 
   const hashPassword = await passwordCryptr(credential.password);
-  console.log("aqui", credential.password);
   await credentialRepository.createCredential(userId, {
     ...credential,
     password: hashPassword,
@@ -27,10 +30,17 @@ async function showCredentials(userId: number) {
 
 async function showCredentialById(id: number, userId: number) {
   const credentialById = await credentialRepository.getCredential(id, userId);
+  if (!credentialById) {
+    throw notFoundError();
+  }
   return credentialById;
 }
 async function deleteCredential(id: number, userId: number) {
-  return await credentialRepository.deleteCredential(id);
+  const credentialCheck = await showCredentialById(id, userId);
+  if (!credentialCheck) {
+    throw notFoundError();
+  }
+  await credentialRepository.deleteCredential(id);
 }
 const credentialService = {
   newCredential,
